+++ gulpfile.js
@@ -350,6 +350,7 @@ async function runTests(cover, all) {
 	try {
 		process.env.APPCD_TEST_GLOBAL_PACKAGE_DIR = path.join(__dirname, 'packages');
 		process.env.SPAWN_WRAP_SHIM_ROOT = origHomeDir;
+		process.env.SNOOPLOGG = '*';
 
 		tmpHomeDir = tmp.dirSync({
 			mode: '755',
@@ -381,6 +382,7 @@ exports['functional-test']      = series(nodeInfo, build, function test() {     
 exports['functional-test-only'] = series(nodeInfo,        function test() {     return runTests(); });
 exports['functional-coverage']  = series(nodeInfo,        function coverage() { return runTests(true); });
 exports['coverage']             = series(nodeInfo,        function coverage() { return runTests(true, true); });
+exports['coverage-ci']          = series(nodeInfo, build, function coverage() { return runTests(true, true); });
 
 /*
  * watch/debug tasks